pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[project]
name = "quantus"
version = "0.0.1"
authors = [
    { name = "David Spector", email = "davidspector9@gmail.com" }
]

description = "Quantitative Ultrasound for Python"
readme = "README.md"
requires-python = ">=3.10.0"

classifiers = [
    "Development Status :: 4 - Beta",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Topic :: Scientific/Engineering",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Software Development",
    "Topic :: Software Development :: Libraries",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Intended Audience :: Developers",
    "Intended Audience :: Education",
    "Intended Audience :: Science/Research",
    "Operating System :: OS Independent",
]
dependencies = [
    "appdirs>=1.4.4",
    "contourpy>=1.1.0",
    "cycler>=0.11.0",
    "et-xmlfile>=1.1.0",
    "fonttools>=4.42.1",
    "gdown>=5.1.0",
    "h5py>=3.9.0",
    "imageio>=2.31.1",
    "joblib>=1.3.2",
    "kiwisolver>=1.4.5",
    "mat73>=0.60",
    "matplotlib>=3.6.2",
    "matplotlib-inline>=0.1.3",
    "networkx>=3.1",
    "numpy>=1.23.0",
    "opencv-python-headless>=4.8.0.76",
    "openpyxl>=3.0.10",
    "packaging>=23.1",
    "pandas>=1.4.3",
    "Pillow>=9.2.0",
    "pydicom>=2.4.1",
    "pyparsing>=3.1.1",
    "python-dateutil>=2.8.2",
    "pytz>=2023.3",
    "pyvista>=0.36.1",
    "PyWavelets>=1.4.1",
    "PyYAML>=6.0.2",
    "scikit-image>=0.19.3",
    "scikit-learn>=1.1.1",
    "scipy>=1.11.3",
    "scooby>=0.7.2",
    "six>=1.16.0",
    "scikit-learn",
    "threadpoolctl>=3.2.0",
    "tifffile>=2023.8.12",
    "tqdm>=4.66.2",
    "traitlets>=5.9.0",
    "typing_extensions>=4.11.0",
    "vtk>=9.2.6"
]

[project.urls]
"Source" = "https://github.com/TUL-Dev/QuantUS-Plugins"

[project.scripts]
quantus = "quantus.full_workflow:main_yaml"
quantus_cli = "quantus.cli:main_cli"

[tool.setuptools]
ext-modules = [
  {name = "philipsRfParser", sources = ["quantus/image_loading/utc_loaders/philips_rf_sc_3d/philips_rf_parser.c"]},
]

[tool.setuptools.packages.find]
where = ["."]
include = ["quantus*"]
